Split Speedhive parsers and restore Node eslint env for Codacy
Break HTML/Puppeteer scrape helpers into small modules under the LOC and
complexity thresholds; mark newsletter.js as Node so require is defined.
